FairSight/backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
fairlearn
google-genai
numpy
python-multipart
